+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - focus on reveal
+- refresh tree on save/delete
 
 ## [0.3.2] - 2021-05-18
 - yes/no are booleans (https://github.com/kubernetes/kubernetes/issues/34146)